+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.9.0]
+
+### Added
+- Support for API authentication via [Earthdata Login bearer tokens](https://urs.earthdata.nasa.gov/documentation/for_users/user_token)
+
 ## [10.8.0]
 
 ### Added